requirements.txt
gensim
google-api-python-client
nltk
numpy
python-dateutil
scikit-learn
tensorflow-gpu>=0.12.1
